ENGR00219006 cts: disallow local file access
CTS 4.0.3_r3 added one case of:
android.security.cts.BrowserTest -- testBrowserPrivateDataAccess
This case always fail on our platform, which is used
to verfiy the js can access local file or not.

The formal fix is in 4.1:
http://code.google.com/p/android/issues/detail?id=32511

But for us it's too hard to merge it's fix as we still
stay in 4.0.3. Here just put a workaround.
The saved offline page can be loaded with this workaround,
looks like no side-effect mentioned in the issue info.

Signed-off-by: Chen Xinyu <xinyu.chen@freescale.com>